swap components of photo and video cards
