+++ src/components/Home/PhotoCards.jsx
@@ -1,17 +1,5 @@
-import { Swiper, SwiperSlide } from "swiper/react";
+import { Stack } from "../";
 
-import "swiper/css";
-import "swiper/css/effect-coverflow";
-import "swiper/css/pagination";
-import "swiper/css/navigation";
-import {
-  EffectCoverflow,
-  Pagination,
-  Mousewheel,
-  Navigation,
-} from "swiper/modules";
-
-import "../../styles/swiperPhoto.css";
 import slide_image_1 from "/assets/images/product1.jpg";
 import slide_image_2 from "/assets/images/food1.jpg";
 import slide_image_3 from "/assets/images/portrait1.jpg";
@@ -19,118 +7,52 @@ import slide_image_4 from "/assets/images/aesthetic1.jpg";
 import slide_image_5 from "/assets/images/nature1.jpg";
 
 function PhotoCards() {
+  const images = [
+    {
+      id: 1,
+      src: slide_image_1,
+      type: "image",
+      label: "Product",
+    },
+    {
+      id: 2,
+      src: slide_image_5,
+      type: "image",
+      label: "Nature",
+    },
+    {
+      id: 3,
+      src: slide_image_2,
+      type: "image",
+      label: "Food",
+    },
+    {
+      id: 4,
+      src: slide_image_4,
+      type: "image",
+      label: "Aesthetics",
+    },
+    {
+      id: 5,
+      src: slide_image_3,
+      type: "image",
+      label: "Portrait",
+    },
+  ];
+
   return (
     <div className="bg-pallet4 pb-28" id="photos">
-      <div className="container md:px-4 px-0 mx-auto">
-        <div className="md:text-5xl text-4xl font-medium  text-center py-10">
-          Photo Categories
-        </div>
-        <Swiper
-          effect={"coverflow"}
-          grabCursor={true}
-          centeredSlides={true}
-          slidesPerView={1}
-          freeMode={true}
-          loop={true}
-          mousewheel={{
-            forceToAxis: true,
-            realaseOnEdges: true,
-            sensitivity: 0.1,
-            thresholdDelta: 1,
-            thresholdTime: 1000,
-          }}
-          breakpoints={{
-            // Adjust slidesPerView based on screen width
-            640: {
-              slidesPerView: 1,
-            },
-            768: {
-              slidesPerView: 2,
-            },
-            1024: {
-              slidesPerView: 3,
-            },
-          }}
-          speed={800}
-          coverflowEffect={{
-            rotate: 50,
-            stretch: 0,
-            depth: 100,
-            modifier: 1,
-            slideShadows: true,
-          }}
-          pagination={true}
-          modules={[EffectCoverflow, Mousewheel, Pagination, Navigation]}
-          navigation={{
-            nextEl: ".swiper-button-next",
-            prevEl: ".swiper-button-prev",
-          }}
-          className="mySwiperPhoto"
-        >
-          <SwiperSlide>
-            <div className="aspect-ratio-4-5">
-              <img
-                src={slide_image_1}
-                alt="slide_image"
-                className="object-cover w-full h-full"
-              />
-            </div>
-            <div className="absolute font-confine bottom-5 md:left-5 left-8 lg:text-6xl text-4xl font-bold text-pallet4">
-              Product
-            </div>
-          </SwiperSlide>
-          <SwiperSlide>
-            <div className="aspect-ratio-4-5">
-              <img
-                src={slide_image_5}
-                alt="slide_image"
-                className="object-cover w-full h-full"
-              />
-            </div>
-            <div className="absolute font-confine bottom-5 md:left-5 left-8 lg:text-6xl text-4xl font-bold text-pallet4">
-              Nature
-            </div>
-          </SwiperSlide>
-          <SwiperSlide>
-            <div className="aspect-ratio-4-5">
-              <img
-                src={slide_image_2}
-                alt="slide_image"
-                className="object-cover w-full h-full"
-              />
-            </div>
-            <div className="absolute font-confine bottom-5 md:left-5 left-8 lg:text-6xl text-4xl font-bold text-pallet4">
-              Food
-            </div>
-          </SwiperSlide>
-          <SwiperSlide>
-            <div className="aspect-ratio-4-5">
-              <img
-                src={slide_image_4}
-                alt="slide_image"
-                className="object-cover w-full h-full"
-              />
-            </div>
-            <div className="absolute font-confine bottom-5 md:left-5 left-8 lg:text-6xl text-4xl font-bold text-pallet4">
-              Aesthetics
-            </div>
-          </SwiperSlide>
-          <SwiperSlide>
-            <div className="aspect-ratio-4-5">
-              <img
-                src={slide_image_3}
-                alt="slide_image"
-                className="object-cover w-full h-full"
-              />
-            </div>
-            <div className="absolute font-confine bottom-5 md:left-5 left-8 lg:text-6xl text-4xl font-bold text-pallet4">
-              Portrait
-            </div>
-          </SwiperSlide>
-
-          <div className="swiper-button-prev cursor-pointer bg-pallet3 p-10 rounded-full text-white hidden md:flex"></div>
-          <div className="swiper-button-next cursor-pointer bg-pallet3 p-10 rounded-full text-white hidden md:flex"></div>
-        </Swiper>
+      <div className="md:text-5xl text-4xl font-medium text-center py-10">
+        Photo Categories
+      </div>
+      <div className="w-screen flex justify-center">
+        <Stack
+          randomRotation={false}
+          sensitivity={180}
+          sendToBackOnClick={true}
+          cardDimensions={{ width: 800, height: 800 }}
+          cardsData={images}
+        />
       </div>
     </div>
   );
